Split tictactoe files and types
