+++ game-pipeline/TicTacToe/GameEngine.hs
@@ -7,6 +7,7 @@ import System.IO
 import Data.Aeson
 import Data.Aeson.Types (Parser)
 import Types
+import TicTacToe.Types
 import GHC.Generics
 import qualified Data.ByteString.Char8 as B
 import qualified Data.ByteString.Lazy as LB
